algun cambio random de salto de linea
