+++ src/investments/types/plazoFijo.type.ts
@@ -24,29 +24,3 @@ export interface PlazoFijoPaginatedResponseInterface {
   errors: any;
   timestamp: string;
 }
-
-// Tipos para filtros
-export const FILTER_VALUES = {
-  IN_PROGRESS: "IN_PROGRESS",
-  FINISHED: "FINISHED",
-} as const;
-
-export const statusFilters: Array<{
-  value: FIXED_TERM_STATES;
-  label: string;
-  color: string;
-  filterValue: string | null;
-}> = [
-  {
-    value: "IN_PROGRESS",
-    label: "En Progreso",
-    color: "#22c55e",
-    filterValue: FILTER_VALUES.IN_PROGRESS,
-  },
-  {
-    value: "FINISHED",
-    label: "Finalizados",
-    color: "#f59e0b",
-    filterValue: FILTER_VALUES.FINISHED,
-  },
-];
